Flatten the static modal style
